CubeSat-Telemetry-Decoder/requirements.txt
numpy>=1.24
scipy>=1.10
pyrtlsdr>=0.3.0
pyyaml>=6.0
click>=8.1
flask>=3.0        # Ground station web UI
flask-socketio>=5.3
pytest>=7.4
sgp4>=2.23        # Orbit Propagation
skyfield>=1.46    # Astrometry and Passes
paho-mqtt>=1.6.1  # Federation
scikit-learn>=1.3 # Anomaly Detection baselines
ntplib>=0.4.0     # Strict NTP timing
cryptography>=41.0.0 # ECDSA PKI Security
